+++ kube/src/api/resource.rs
@@ -1,6 +1,6 @@
+use super::params::{DeleteParams, ListParams, PatchParams, PostParams};
 use crate::{api::DynamicResource, Error, Result};
 use inflector::string::pluralize::to_plural;
-use serde::Serialize;
 
 /// A Kubernetes resource that can be accessed through the API
 #[derive(Clone, Debug)]
@@ -88,254 +88,6 @@ impl Resource {
     }
 }
 
-/// Common query parameters used in watch/list/delete calls on collections
-#[derive(Default, Clone)]
-#[allow(missing_docs)]
-pub struct ListParams {
-    pub field_selector: Option<String>,
-    pub include_uninitialized: bool,
-    pub label_selector: Option<String>,
-    pub timeout: Option<u32>,
-    pub allow_bookmarks: bool,
-}
-
-impl ListParams {
-    fn validate(&self) -> Result<()> {
-        if let Some(to) = &self.timeout {
-            // https://github.com/kubernetes/kubernetes/issues/6513
-            if *to >= 295 {
-                return Err(Error::RequestValidation(
-                    "ListParams::timeout must be < 295s".into(),
-                ));
-            }
-        }
-        Ok(())
-    }
-}
-
-/// Builder interface to ListParams
-///
-/// Usage:
-/// ```
-/// use kube::api::ListParams;
-/// let lp = ListParams::default()
-///     .timeout(60)
-///     .labels("kubernetes.io/lifecycle=spot");
-/// ```
-impl ListParams {
-    /// Configure the timeout for list/watch calls
-    ///
-    /// This limits the duration of the call, regardless of any activity or inactivity.
-    /// Defaults to 290s
-    pub fn timeout(mut self, timeout_secs: u32) -> Self {
-        self.timeout = Some(timeout_secs);
-        self
-    }
-
-    /// Configure the selector to restrict the list of returned objects by their fields.
-    ///
-    /// Defaults to everything.
-    /// Supports '=', '==', and '!=', and can comma separate: key1=value1,key2=value2
-    /// The server only supports a limited number of field queries per type.
-    pub fn fields(mut self, field_selector: &str) -> Self {
-        self.field_selector = Some(field_selector.to_string());
-        self
-    }
-
-    /// Configure the selector to restrict the list of returned objects by their labels.
-    ///
-    /// Defaults to everything.
-    /// Supports '=', '==', and '!=', and can comma separate: key1=value1,key2=value2
-    pub fn labels(mut self, label_selector: &str) -> Self {
-        self.label_selector = Some(label_selector.to_string());
-        self
-    }
-
-    /// If called, partially initialized resources are included in watch/list responses.
-    pub fn include_uninitialized(mut self) -> Self {
-        self.include_uninitialized = true;
-        self
-    }
-
-    /// Enables watch bookmarks from the api server if supported
-    pub fn allow_bookmarks(mut self) -> Self {
-        self.allow_bookmarks = true;
-        self
-    }
-}
-
-// TODO: WatchParams (same as ListParams but with extra resource_version + allow_watch_bookmarks)
-
-/// Common query parameters for put/post calls
-#[derive(Default, Clone)]
-pub struct PostParams {
-    /// Whether to run this as a dry run
-    pub dry_run: bool,
-}
-
-/// Common query parameters for patch calls
-#[derive(Default, Clone)]
-pub struct PatchParams {
-    /// Whether to run this as a dry run
-    pub dry_run: bool,
-    /// Strategy which will be used. Defaults to `PatchStrategy::Strategic`
-    pub patch_strategy: PatchStrategy,
-    /// force Apply requests. Applicable only to `PatchStrategy::Apply`
-    pub force: bool,
-    /// fieldManager is a name of the actor that is making changes. Required for `PatchStrategy::Apply`
-    /// optional for everything else
-    pub field_manager: Option<String>,
-}
-impl PatchParams {
-    fn validate(&self) -> Result<()> {
-        if let Some(field_manager) = &self.field_manager {
-            // Implement the easy part of validation, in future this may be extended to provide validation as in go code
-            // For now it's fine, because k8s API server will return an error
-            if field_manager.len() > 128 {
-                return Err(Error::RequestValidation(
-                    "Failed to validate PatchParams::field_manager!".into(),
-                ));
-            }
-        }
-
-        if self.patch_strategy != PatchStrategy::Apply && self.force {
-            // if not force, all other fields are valid for all types of patch requests
-            Err(Error::RequestValidation(
-                "Force is applicable only for Apply strategy!".into(),
-            ))
-        } else {
-            Ok(())
-        }
-    }
-
-    fn populate_qp(&self, qp: &mut url::form_urlencoded::Serializer<String>) {
-        if self.dry_run {
-            qp.append_pair("dryRun", "true");
-        }
-        if self.force {
-            qp.append_pair("force", "true");
-        }
-        if let Some(ref field_manager) = self.field_manager {
-            qp.append_pair("fieldManager", &field_manager);
-        }
-    }
-
-    /// Construct PatchParams for server-side apply
-    pub fn apply(manager: &str) -> Self {
-        Self {
-            patch_strategy: PatchStrategy::Apply,
-            field_manager: Some(manager.into()),
-            ..Self::default()
-        }
-    }
-
-    /// Force the result through on conflicts
-    pub fn force(mut self) -> Self {
-        self.force = true;
-        self
-    }
-
-    /// Perform a dryRun only
-    pub fn dry_run(mut self) -> Self {
-        self.dry_run = true;
-        self
-    }
-}
-
-/// Four different patch types are supported.
-///
-/// Apply strategy is kinda special
-#[derive(Clone, PartialEq)]
-pub enum PatchStrategy {
-    /// [Server side apply](https://kubernetes.io/docs/reference/using-api/api-concepts/#server-side-apply)
-    ///
-    /// Requires kubernetes >=1.16
-    Apply,
-    /// A [JSON merge](https://kubernetes.io/docs/tasks/run-application/update-api-object-kubectl-patch/#use-a-json-merge-patch-to-update-a-deployment)
-    JSON,
-    /// A regular merge
-    Merge,
-    /// A [stategic merge](https://kubernetes.io/docs/tasks/run-application/update-api-object-kubectl-patch/#use-a-strategic-merge-patch-to-update-a-deployment)
-    Strategic,
-}
-
-impl std::fmt::Display for PatchStrategy {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let content_type = match &self {
-            Self::Apply => "application/apply-patch+yaml",
-            Self::JSON => "application/json-patch+json",
-            Self::Merge => "application/merge-patch+json",
-            Self::Strategic => "application/strategic-merge-patch+json",
-        };
-        f.write_str(content_type)
-    }
-}
-
-// Kubectl defaults to Strategic strategy, but doing so will break existing consumers
-// so, currently we still default to Merge it may change in future versions
-// Strategic merge doesn't work with CRD types https://github.com/kubernetes/kubernetes/issues/52772
-impl Default for PatchStrategy {
-    fn default() -> Self {
-        PatchStrategy::Merge
-    }
-}
-
-/// Common query parameters for delete calls
-#[derive(Default, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DeleteParams {
-    /// When present, indicates that modifications should not be persisted.
-    ///
-    /// An invalid or unrecognized dryRun directive will result in an error response
-    /// and no further processing of the request.
-    /// Valid values are:
-    /// - All: all dry run stages will be processed
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub dry_run: Vec<String>,
-
-    /// The duration in seconds before the object should be deleted.
-    ///
-    /// Value must be non-negative integer. The value zero indicates delete immediately.
-    /// If this value is None, the default grace period for the specified type will be used.
-    /// Defaults to a per object value if not specified. Zero means delete immediately.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub grace_period_seconds: Option<u32>,
-
-    /// Whether or how garbage collection is performed.
-    ///
-    /// The default policy is decided by the existing finalizer set in
-    /// metadata.finalizers, and the resource-specific default policy.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub propagation_policy: Option<PropagationPolicy>,
-
-    /// Condtions that must be fulfilled before a deletion is carried out
-    ///
-    /// If not possible, a 409 Conflict status will be returned.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub preconditions: Option<Preconditions>,
-}
-
-/// Preconditions must be fulfilled before an operation (update, delete, etc.) is carried out.
-#[derive(Default, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Preconditions {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub resource_version: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub uid: Option<String>,
-}
-
-/// Propagation policy when deleting single objects
-#[derive(Clone, Debug, Serialize)]
-pub enum PropagationPolicy {
-    /// Orphan dependents
-    Orphan,
-    /// Allow the garbage collector to delete the dependents in the background
-    Background,
-    /// A cascading policy that deletes all dependents in the foreground
-    Foreground,
-}
-
 /// Convenience methods found from API conventions
 impl Resource {
     /// List a collection of a resource
@@ -345,9 +97,6 @@ impl Resource {
 
         if let Some(fields) = &lp.field_selector {
             qp.append_pair("fieldSelector", &fields);
-        }
-        if lp.include_uninitialized {
-            qp.append_pair("includeUninitialized", "true");
         }
         if let Some(labels) = &lp.label_selector {
             qp.append_pair("labelSelector", &labels);
@@ -372,9 +121,6 @@ impl Resource {
         if let Some(fields) = &lp.field_selector {
             qp.append_pair("fieldSelector", &fields);
         }
-        if lp.include_uninitialized {
-            qp.append_pair("includeUninitialized", "true");
-        }
         if let Some(labels) = &lp.label_selector {
             qp.append_pair("labelSelector", &labels);
         }
@@ -398,6 +144,7 @@ impl Resource {
 
     /// Create an instance of a resource
     pub fn create(&self, pp: &PostParams, data: Vec<u8>) -> Result<http::Request<Vec<u8>>> {
+        pp.validate()?;
         let base_url = self.make_url() + "?";
         let mut qp = url::form_urlencoded::Serializer::new(base_url);
         if pp.dry_run {
@@ -424,9 +171,6 @@ impl Resource {
         let mut qp = url::form_urlencoded::Serializer::new(base_url);
         if let Some(fields) = &lp.field_selector {
             qp.append_pair("fieldSelector", &fields);
-        }
-        if lp.include_uninitialized {
-            qp.append_pair("includeUninitialized", "true");
         }
         if let Some(labels) = &lp.label_selector {
             qp.append_pair("labelSelector", &labels);